Moved VersionInfoMessage, DeviceIDMessage, and EventNotificationMessage to device.h.
These are informative status messages, not commands/control messages.
